pyproject.toml
[project]
name = "eye"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi>=0.115.12",
    "numpy>=2.1.3",
    "openai>=1.76.0",
    "opencv-python>=4.11.0.86",
    "passlib>=1.7.4",
    "pydantic>=2.11.3",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.40",
    "tensorflow>=2.19.0",
    "uvicorn[standard]>=0.34.2",
]
